requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.9
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-debug-toolbar==3.2.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
model-bakery==1.3.2
msgpack==1.0.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylibmc==1.6.1
pymemcache==3.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
